ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
nltk==3.8.1
python-dotenv==1.0.0
pymongo==4.6.0
pyjwt==2.8.0
